refactor(workers): Move auth-related classes to a separate package
There are already a number of classes dealing with authentication, and
this is going to be extended. Therefore, having a dedicated package for
authentication makes sense. Also, these classes should not be tightly
connected to the worker context.
